NBSN95 v1.2.7
v1.2.7
1.Compatible with LDDS45 distance sensor

v1.2.6
1.Compatible with BC95-G, BC35-G and BG95-GV modules

+++ NBSN95/Drivers/BSP/inc/time_server.h
@@ -31,7 +31,7 @@ typedef struct TimerEvent_s
     struct TimerEvent_s *Next;  //! Pointer to the next Timer object.
 } TimerEvent_t;
 
-void My_AlarmInit(uint16_t timer,uint8_t alarmX);
+void My_AlarmInit(uint32_t timer,uint8_t alarmX);
 void MyRtcInit(void);
 long GetTick(char *str_time);
 #endif
